refactor: add assistant feature entrypoint
