+++ src/app/(home)/blog/[slug]/page.tsx
@@ -5,7 +5,7 @@ import { ShareMenu } from '@/app/(home)/blog/[slug]/page.client'
 import { mdxComponents } from '@/components/content/mdx/components'
 import { GitHubCode } from '@/components/content/mdx/github-code'
 import { MdxContent } from '@/components/content/mdx/layout'
-import { LLMCopyButtonWithViewOptions } from '@/components/features/assistant/page-actions'
+import { LLMCopyButtonWithViewOptions } from '@/components/features/assistant'
 import { TOCMinimap, TOCPopover } from '@/components/features/toc'
 import { PostJsonLd } from '@/components/json-ld'
 import { SectionBody } from '@/components/layout/sections/body'
