+++ src/app/provider.tsx
@@ -5,8 +5,7 @@ import { SoundProvider } from '@web-kits/audio/react'
 import { NuqsAdapter } from 'nuqs/adapters/next/app'
 import type { ReactNode } from 'react'
 import Analytics from '@/components/analytics'
-import { AISearch } from '@/components/features/assistant/chat'
-import { AISearchTrigger } from '@/components/features/assistant/search-trigger'
+import { AISearch, AISearchTrigger } from '@/components/features/assistant'
 import { MobileNav } from '@/components/layout/header/mobile'
 import { SmoothCursor } from '@/components/smooth-cursor'
 import { TailwindIndicator } from '@/components/tailwind-indicator'
