+++ src/components/features/assistant/index.ts
@@ -1,0 +1,12 @@
+export {
+  AISearch,
+  useAISearchContext,
+  useChatContext,
+} from './chat'
+export {
+  LLMCopyButton,
+  LLMCopyButtonWithViewOptions,
+  ViewOptions,
+} from './page-actions'
+export { AISearchTrigger } from './search-trigger'
+export { SelectionContent } from './selection-context-menu'
